Reconfigure article view
